requirements.txt
asgiref==3.8.1
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
contourpy==1.3.0
cycler==0.12.1
Django==4.2.18
django-ranged-response==0.2.0
django-simple-captcha==0.6.1
djangorestframework==3.15.2
fonttools==4.55.8
frozendict==2.4.6
gunicorn==23.0.0
html5lib==1.1
idna==3.10
importlib_resources==6.5.2
Jinja2==3.1.5
kiwisolver==1.4.7
lxml==5.3.0
MarkupSafe==3.0.2
matplotlib==3.9.4
multitasking==0.0.11
numpy==2.0.2
packaging==24.2
pandas==2.2.3
peewee==3.17.9
pillow==11.1.0
platformdirs==4.3.6
prettytable==3.14.0
pyecharts==2.0.8
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2025.1
requests==2.32.3
simplejson==3.19.3
six==1.17.0
soupsieve==2.6
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
wcwidth==0.2.13
webencodings==0.5.1
yfinance==0.2.52
zipp==3.21.0
